move shared drawer spinner to its own component
